down to 9 errors
It seems all the errors are stemming from the triangle struct. Other header files and classes are unable to access it and the std::vector class doesn't know how to parse the information explicitly. I will have to dig into the original olcEngine header file and see if there was a hack-fix in there.

Bedtime.